+++ R/SCD-effect-sizes.R
@@ -12,7 +12,7 @@
 
 SCD_effect_sizes <- function(browser = TRUE) {
   
-  req_pkgs <- c("shiny", "ggplot2", "markdown", "rmarkdown","tidyr", "dplyr", "readxl", 
+  req_pkgs <- c("shiny", "ggplot2", "markdown", "tidyr", "dplyr", "readxl", 
                 "janitor", "rlang", "glue", "rclipboard", "Kendall")
   missing_pkgs <- unlist(lapply(req_pkgs, check_for_package))